indentation and files rename fixes
